Edit Post: Split Block Manager category render to own component
